feat: Rename Proposal to AgenticRun with backward compatibility
Aligns with upstream CRD rename in lightspeed-agentic-operator.
All internal code now uses AgenticRun naming, while maintaining
full backward compatibility for existing configurations.

Changes:
- Core: AgenticRunDriver, AgenticRunAmender, AgenticRunAgentConfig
- Models: agentic_run_spec, agentic_run_status, agentic_run_results
- Metrics: custom:agentic_run_status, custom:agentic_run_evaluation_correctness
- Backward compat via Pydantic AliasChoices and @property accessors
- Deprecation warnings for old "proposal" names
- Integration tests updated and namespace corrected

All 1399 unit tests passing, quality checks clean.
