+++ src/lightspeed_evaluation/core/agentic_run/__init__.py
@@ -1,0 +1,5 @@
+"""AgenticRun CRD domain logic."""
+
+from lightspeed_evaluation.core.agentic_run.phase import derive_phase
+
+__all__ = ["derive_phase"]
