+++ src/lightspeed_evaluation/core/metrics/custom/__init__.py
@@ -1,5 +1,8 @@
 """Custom metrics components package."""
 
+from lightspeed_evaluation.core.metrics.custom.agentic_run_eval import (
+    evaluate_agentic_run_status,
+)
 from lightspeed_evaluation.core.metrics.custom.conformal import (
     compute_mrr_threshold,
     get_lhat,
@@ -12,16 +15,17 @@ from lightspeed_evaluation.core.metrics.custom.prompts import (
     INTENT_EVALUATION_PROMPT,
 )
 from lightspeed_evaluation.core.metrics.custom.proposal_eval import (
-    evaluate_proposal_status,
+    evaluate_proposal_status,  # Deprecated alias
 )
 from lightspeed_evaluation.core.metrics.custom.tool_eval import evaluate_tool_calls
 
 __all__ = [
     "CustomMetrics",
     "compute_mrr_threshold",
+    "evaluate_agentic_run_status",
     "evaluate_keywords",
     "evaluate_mrr",
-    "evaluate_proposal_status",
+    "evaluate_proposal_status",  # Deprecated alias
     "evaluate_tool_calls",
     "get_lhat",
     # Prompts
