+++ src/lightspeed_evaluation/core/models/data.py
@@ -5,6 +5,7 @@ from pathlib import Path
 from typing import Annotated, Any, Optional
 
 from pydantic import (
+    AliasChoices,
     BaseModel,
     BeforeValidator,
     ConfigDict,
@@ -142,13 +143,13 @@ def _validate_and_deduplicate_metrics(
 class TurnData(StreamingMetricsMixin):
     """Individual turn data within a conversation."""
 
-    model_config = ConfigDict(extra="forbid")
+    model_config = ConfigDict(extra="forbid", populate_by_name=True)
 
     turn_id: str = Field(..., min_length=1, description="Turn ID (alphanumeric)")
     query: Optional[str] = Field(
         default=None,
         min_length=1,
-        description="Query — auto-populated from proposal_spec.request when absent",
+        description="Query — auto-populated from agentic_run_spec.request when absent",
     )
     attachments: Optional[list[str]] = Field(
         default=None, min_length=0, description="Attachments"
@@ -239,26 +240,35 @@ class TurnData(StreamingMetricsMixin):
         default=None, description="Path to verify script for script-based evaluation"
     )
 
-    # Proposal driver fields
+    # AgenticRun driver fields
     description: Optional[str] = Field(
         default=None, description="Human-readable label for reports"
     )
-    proposal_spec: Optional[dict[str, Any]] = Field(
-        default=None, description="Inline proposal spec for CRD-based agents"
-    )
-    expected_proposal_status: Optional[dict[str, Any]] = Field(
+    agentic_run_spec: Optional[dict[str, Any]] = Field(
         default=None,
-        description="Expected proposal status for assertion metrics",
+        validation_alias=AliasChoices("agentic_run_spec", "proposal_spec"),
+        description="Inline agentic run spec for CRD-based agents",
     )
-    proposal_status: Optional[dict[str, Any]] = Field(
-        default=None, description="Raw CRD status populated by ProposalDriver"
-    )
-    proposal_results: Optional[dict[str, Any]] = Field(
+    expected_agentic_run_status: Optional[dict[str, Any]] = Field(
         default=None,
-        description="Structured results from child Result CRs, populated by ProposalAmender",
+        validation_alias=AliasChoices(
+            "expected_agentic_run_status", "expected_proposal_status"
+        ),
+        description="Expected agentic run status for assertion metrics",
     )
-    proposal_phases: Optional[list[str]] = Field(
+    agentic_run_status: Optional[dict[str, Any]] = Field(
         default=None,
+        validation_alias=AliasChoices("agentic_run_status", "proposal_status"),
+        description="Raw CRD status populated by AgenticRunDriver",
+    )
+    agentic_run_results: Optional[dict[str, Any]] = Field(
+        default=None,
+        validation_alias=AliasChoices("agentic_run_results", "proposal_results"),
+        description="Structured results from child Result CRs, populated by AgenticRunAmender",
+    )
+    agentic_run_phases: Optional[list[str]] = Field(
+        default=None,
+        validation_alias=AliasChoices("agentic_run_phases", "proposal_phases"),
         description="Workflow phases that actually executed (e.g. ['analysis', 'execution'])",
     )
 
@@ -273,21 +283,72 @@ class TurnData(StreamingMetricsMixin):
         """Returns True if the metric didn't pass the validation."""
         return metric in self._invalid_metrics
 
+    # Deprecated property accessors for backward compatibility
+    @property
+    def proposal_spec(self) -> Optional[dict[str, Any]]:
+        """Deprecated: use agentic_run_spec instead."""
+        return self.agentic_run_spec
+
+    @proposal_spec.setter
+    def proposal_spec(self, value: Optional[dict[str, Any]]) -> None:
+        """Deprecated setter for proposal_spec."""
+        self.agentic_run_spec = value
+
+    @property
+    def expected_proposal_status(self) -> Optional[dict[str, Any]]:
+        """Deprecated: use expected_agentic_run_status instead."""
+        return self.expected_agentic_run_status
+
+    @expected_proposal_status.setter
+    def expected_proposal_status(self, value: Optional[dict[str, Any]]) -> None:
+        """Deprecated setter for expected_proposal_status."""
+        self.expected_agentic_run_status = value
+
+    @property
+    def proposal_status(self) -> Optional[dict[str, Any]]:
+        """Deprecated: use agentic_run_status instead."""
+        return self.agentic_run_status
+
+    @proposal_status.setter
+    def proposal_status(self, value: Optional[dict[str, Any]]) -> None:
+        """Deprecated setter for proposal_status."""
+        self.agentic_run_status = value
+
+    @property
+    def proposal_results(self) -> Optional[dict[str, Any]]:
+        """Deprecated: use agentic_run_results instead."""
+        return self.agentic_run_results
+
+    @proposal_results.setter
+    def proposal_results(self, value: Optional[dict[str, Any]]) -> None:
+        """Deprecated setter for proposal_results."""
+        self.agentic_run_results = value
+
+    @property
+    def proposal_phases(self) -> Optional[list[str]]:
+        """Deprecated: use agentic_run_phases instead."""
+        return self.agentic_run_phases
+
+    @proposal_phases.setter
+    def proposal_phases(self, value: Optional[list[str]]) -> None:
+        """Deprecated setter for proposal_phases."""
+        self.agentic_run_phases = value
+
     @model_validator(mode="after")
-    def populate_query_from_proposal_spec(self) -> "TurnData":
-        """Auto-populate query from proposal_spec.request when absent."""
+    def populate_query_from_agentic_run_spec(self) -> "TurnData":
+        """Auto-populate query from agentic_run_spec.request when absent."""
         if self.query is not None:
             return self
-        if self.proposal_spec is not None:
-            request = self.proposal_spec.get("request")
+        if self.agentic_run_spec is not None:
+            request = self.agentic_run_spec.get("request")
             if isinstance(request, str) and request.strip():
                 self.query = request
                 return self
             raise ValueError(
-                "proposal_spec must contain a non-empty 'request' "
+                "agentic_run_spec must contain a non-empty 'request' "
                 "when query is not provided"
             )
-        raise ValueError("query is required when proposal_spec is not provided")
+        raise ValueError("query is required when agentic_run_spec is not provided")
 
     @field_validator("turn_metrics")
     @classmethod
