+++ src/lightspeed_evaluation/core/proposal/__init__.py
@@ -1,5 +1,18 @@
-"""Proposal CRD domain logic."""
+"""Deprecated: Use core.agentic_run instead.
 
-from lightspeed_evaluation.core.proposal.phase import derive_phase
+This module is deprecated and maintained for backward compatibility only.
+Use lightspeed_evaluation.core.agentic_run instead.
+"""
+
+import logging
+
+# Re-export from agentic_run for backward compatibility
+from lightspeed_evaluation.core.agentic_run import derive_phase
 
 __all__ = ["derive_phase"]
+
+logger = logging.getLogger(__name__)
+logger.warning(
+    "Module 'lightspeed_evaluation.core.proposal' is deprecated. "
+    "Use 'lightspeed_evaluation.core.agentic_run' instead."
+)
