+++ src/lightspeed_evaluation/pipeline/evaluation/driver.py
@@ -12,20 +12,20 @@ from abc import ABC, abstractmethod
 from enum import StrEnum
 from typing import Any, Optional, cast
 
+from lightspeed_evaluation.core.agentic_run import derive_phase
 from lightspeed_evaluation.core.api import APIClient
 from lightspeed_evaluation.core.models import (
+    AgenticRunAgentConfig,
     APIConfig,
     HttpApiAgentConfig,
-    ProposalAgentConfig,
     TurnData,
 )
-from lightspeed_evaluation.core.proposal import derive_phase
 from lightspeed_evaluation.core.system.exceptions import ConfigurationError
+from lightspeed_evaluation.pipeline.evaluation.agentic_run_amender import (
+    AgenticRunAmender,
+)
 from lightspeed_evaluation.pipeline.evaluation.amender import APIDataAmender
 from lightspeed_evaluation.pipeline.evaluation.cli import KubeCLI
-from lightspeed_evaluation.pipeline.evaluation.proposal_amender import (
-    ProposalAmender,
-)
 
 logger = logging.getLogger(__name__)
 
@@ -101,12 +101,12 @@ class HttpApiDriver(AgentDriver):
 
 
 # ---------------------------------------------------------------------------
-# Proposal driver — CRD-based agent lifecycle via oc/kubectl CLI
+# AgenticRun driver — CRD-based agent lifecycle via oc/kubectl CLI
 # ---------------------------------------------------------------------------
 
 
 class TerminalOutcome(StrEnum):
-    """Terminal outcomes for a Proposal CR lifecycle.
+    """Terminal outcomes for an AgenticRun CR lifecycle.
 
     These are driver-level labels, not CRD API values. The CRD exposes
     conditions with statuses that the driver interprets:
@@ -133,11 +133,11 @@ CRD_PLURAL = "agenticruns"
 CRD_API_VERSION = f"{CRD_GROUP}/{CRD_VERSION}"
 
 
-class ProposalDriver(AgentDriver):
-    """Driver that manages Proposal CR lifecycle via oc/kubectl CLI."""
+class AgenticRunDriver(AgentDriver):
+    """Driver that manages AgenticRun CR lifecycle via oc/kubectl CLI."""
 
     def __init__(self, config: dict[str, Any], *, enabled: bool = True) -> None:
-        """Initialize the proposal driver."""
+        """Initialize the agentic run driver."""
         super().__init__(config, enabled=enabled)
         self._cli = self._resolve_cli()
         self._kube_cli = KubeCLI(
@@ -145,11 +145,11 @@ class ProposalDriver(AgentDriver):
             namespace=self._config.namespace,
             timeout=self._config.cli_timeout,
         )
-        self._amender = ProposalAmender(self._kube_cli)
+        self._amender = AgenticRunAmender(self._kube_cli)
 
-    def validate_config(self, config: dict[str, Any]) -> ProposalAgentConfig:
-        """Validate proposal driver configuration."""
-        parsed = ProposalAgentConfig.model_validate(config)
+    def validate_config(self, config: dict[str, Any]) -> AgenticRunAgentConfig:
+        """Validate agentic run driver configuration."""
+        parsed = AgenticRunAgentConfig.model_validate(config)
         if self._enabled and not shutil.which("oc") and not shutil.which("kubectl"):
             raise ConfigurationError("Neither 'oc' nor 'kubectl' found on PATH")
         return parsed
@@ -157,15 +157,15 @@ class ProposalDriver(AgentDriver):
     def execute_turn(
         self, turn_data: TurnData, conversation_id: Optional[str] = None
     ) -> tuple[Optional[str], Optional[str]]:
-        """Execute a Proposal CR lifecycle for a single turn."""
-        # Proposal CR lifecycle:
-        # 1. Build Proposal CR from TurnData fields
+        """Execute an AgenticRun CR lifecycle for a single turn."""
+        # AgenticRun CR lifecycle:
+        # 1. Build AgenticRun CR from TurnData fields
         # 2. Apply CR to cluster via oc/kubectl
         # 3. Poll status — read .status.conditions each interval
-        # 4. Auto-approve — create ProposalApproval when Analyzed=True
+        # 4. Auto-approve — create AgenticRunApproval when Analyzed=True
         # 5. Terminal detection — break on completed/failed/denied/escalated
-        # 6. Amend TurnData — set response and proposal_status in-place
-        # 7. Cleanup — delete Proposal CR if cleanup_proposals enabled
+        # 6. Amend TurnData — set response and agentic_run_status in-place
+        # 7. Cleanup — delete AgenticRun CR if cleanup_proposals enabled
         suffix = uuid.uuid4().hex[:8]
         safe_id = (
             re.sub(r"[^a-z0-9-]", "", conversation_id.lower())[:50]
@@ -173,8 +173,8 @@ class ProposalDriver(AgentDriver):
             else ""
         )
         cr_name = f"eval-{safe_id}-{suffix}" if safe_id else f"eval-{suffix}"
-        proposal_spec = turn_data.proposal_spec or {}
-        manifest = self._build_proposal_cr(turn_data, cr_name)
+        agentic_run_spec = turn_data.agentic_run_spec or {}
+        manifest = self._build_agentic_run_cr(turn_data, cr_name)
 
         result = self._apply(manifest)
         if result.returncode != 0:
@@ -184,7 +184,7 @@ class ProposalDriver(AgentDriver):
             )
 
         if self._config.auto_approve:
-            err = self._approve_when_ready(cr_name, proposal_spec)
+            err = self._approve_when_ready(cr_name, agentic_run_spec)
             if err:
                 self._cleanup(cr_name)
                 return (err, None)
@@ -201,7 +201,7 @@ class ProposalDriver(AgentDriver):
                 return (err, None)
             conditions = status_dict.get("conditions", [])
 
-            outcome = self._is_terminal(conditions, proposal_spec)
+            outcome = self._is_terminal(conditions, agentic_run_spec)
             if outcome is not None:
                 break
         else:
@@ -213,20 +213,20 @@ class ProposalDriver(AgentDriver):
 
         self._amend_turn_data(turn_data, status_dict)
         self._cleanup(cr_name)
-        logger.info("Proposal '%s' reached terminal state: %s", cr_name, outcome)
+        logger.info("AgenticRun '%s' reached terminal state: %s", cr_name, outcome)
         return (None, None)
 
     def _amend_turn_data(
         self, turn_data: TurnData, status_dict: dict[str, Any]
     ) -> None:
-        """Amend turn data from proposal status, with fallback on amender failure."""
+        """Amend turn data from agentic run status, with fallback on amender failure."""
         amend_err = self._amender.amend(turn_data, status_dict)
         if amend_err:
-            logger.warning("ProposalAmender failed: %s", amend_err)
+            logger.warning("AgenticRunAmender failed: %s", amend_err)
             if not turn_data.response:
                 turn_data.response = self._extract_summary(status_dict)
-            if not turn_data.proposal_status:
-                turn_data.proposal_status = status_dict
+            if not turn_data.agentic_run_status:
+                turn_data.agentic_run_status = status_dict
 
     @staticmethod
     def _resolve_cli() -> str:
@@ -238,31 +238,33 @@ class ProposalDriver(AgentDriver):
         return self._kube_cli.apply(manifest)
 
     def _get_status(self, cr_name: str) -> tuple[dict[str, Any], Optional[str]]:
-        """Get Proposal CR status."""
+        """Get AgenticRun CR status."""
         cr, err = self._kube_cli.get_resource(CRD_PLURAL, cr_name)
         if err:
             return {}, f"Failed to get status for '{cr_name}': {err}"
         return cr.get("status", {}), None
 
     def _delete(self, cr_name: str) -> None:
-        """Delete a Proposal CR."""
+        """Delete an AgenticRun CR."""
         self._kube_cli.delete(CRD_PLURAL, cr_name)
 
     def _cleanup(self, cr_name: str) -> None:
-        """Delete the Proposal CR if cleanup is enabled."""
+        """Delete the AgenticRun CR if cleanup is enabled."""
         if not self._config.cleanup_proposals:
             return
         try:
             self._delete(cr_name)
-            logger.info("Cleaned up Proposal CR '%s'", cr_name)
+            logger.info("Cleaned up AgenticRun CR '%s'", cr_name)
         except (subprocess.SubprocessError, OSError) as exc:
-            logger.warning("Failed to clean up Proposal CR '%s': %s", cr_name, exc)
+            logger.warning("Failed to clean up AgenticRun CR '%s': %s", cr_name, exc)
 
-    def _build_proposal_cr(self, turn_data: TurnData, cr_name: str) -> dict[str, Any]:
-        """Build Proposal CR manifest from TurnData."""
+    def _build_agentic_run_cr(
+        self, turn_data: TurnData, cr_name: str
+    ) -> dict[str, Any]:
+        """Build AgenticRun CR manifest from TurnData."""
         spec: dict[str, Any] = {"request": turn_data.query}
-        if turn_data.proposal_spec:
-            spec.update(turn_data.proposal_spec)
+        if turn_data.agentic_run_spec:
+            spec.update(turn_data.agentic_run_spec)
         spec.setdefault("analysis", {})
         return {
             "apiVersion": CRD_API_VERSION,
@@ -275,12 +277,14 @@ class ProposalDriver(AgentDriver):
         }
 
     def _build_approval_cr(
-        self, cr_name: str, proposal_spec: dict[str, Any]
+        self, cr_name: str, agentic_run_spec: dict[str, Any]
     ) -> dict[str, Any]:
-        """Build ProposalApproval CR manifest."""
+        """Build AgenticRunApproval CR manifest."""
         analysis_params: dict[str, Any] = {}
-        if "analysis" in proposal_spec and isinstance(proposal_spec["analysis"], dict):
-            agent = proposal_spec["analysis"].get("agent")
+        if "analysis" in agentic_run_spec and isinstance(
+            agentic_run_spec["analysis"], dict
+        ):
+            agent = agentic_run_spec["analysis"].get("agent")
             if agent:
                 analysis_params["agent"] = agent
         if not analysis_params:
@@ -289,10 +293,10 @@ class ProposalDriver(AgentDriver):
         stages: list[dict[str, Any]] = [
             {"type": "Analysis", "decision": "Approved", "analysis": analysis_params},
         ]
-        if "execution" in proposal_spec:
+        if "execution" in agentic_run_spec:
             exec_params: dict[str, Any] = {"option": 0}
-            if isinstance(proposal_spec["execution"], dict):
-                agent = proposal_spec["execution"].get("agent")
+            if isinstance(agentic_run_spec["execution"], dict):
+                agent = agentic_run_spec["execution"].get("agent")
                 if agent:
                     exec_params["agent"] = agent
             stages.append(
@@ -302,10 +306,10 @@ class ProposalDriver(AgentDriver):
                     "execution": exec_params,
                 }
             )
-        if "verification" in proposal_spec:
+        if "verification" in agentic_run_spec:
             verif_params: dict[str, Any] = {}
-            if isinstance(proposal_spec["verification"], dict):
-                agent = proposal_spec["verification"].get("agent")
+            if isinstance(agentic_run_spec["verification"], dict):
+                agent = agentic_run_spec["verification"].get("agent")
                 if agent:
                     verif_params["agent"] = agent
             if not verif_params:
@@ -328,9 +332,9 @@ class ProposalDriver(AgentDriver):
         }
 
     def _approve_when_ready(
-        self, cr_name: str, proposal_spec: dict[str, Any]
+        self, cr_name: str, agentic_run_spec: dict[str, Any]
     ) -> Optional[str]:
-        """Wait for Proposal CR to exist on the cluster, then approve all stages."""
+        """Wait for AgenticRun CR to exist on the cluster, then approve all stages."""
         start = time.monotonic()
         while time.monotonic() - start < self._config.timeout:
             _, err = self._get_status(cr_name)
@@ -338,12 +342,12 @@ class ProposalDriver(AgentDriver):
                 break
             time.sleep(self._config.poll_interval)
         else:
-            return f"Proposal '{cr_name}' not found within {self._config.timeout}s"
+            return f"AgenticRun '{cr_name}' not found within {self._config.timeout}s"
 
-        approval = self._build_approval_cr(cr_name, proposal_spec)
+        approval = self._build_approval_cr(cr_name, agentic_run_spec)
         result = self._apply(approval)
         if result.returncode != 0:
-            return f"Failed to apply ProposalApproval: {result.stderr.strip()}"
+            return f"Failed to apply AgenticRunApproval: {result.stderr.strip()}"
         return None
 
     _PHASE_TO_OUTCOME: dict[str, TerminalOutcome] = {
@@ -355,21 +359,25 @@ class ProposalDriver(AgentDriver):
 
     @staticmethod
     def _is_terminal(
-        conditions: list[dict[str, Any]], proposal_spec: dict[str, Any]
+        conditions: list[dict[str, Any]], agentic_run_spec: dict[str, Any]
     ) -> Optional[TerminalOutcome]:
         """Check if conditions indicate a terminal state."""
-        phase = derive_phase(conditions, proposal_spec or None)
-        return ProposalDriver._PHASE_TO_OUTCOME.get(phase)
+        phase = derive_phase(conditions, agentic_run_spec=agentic_run_spec or None)
+        return AgenticRunDriver._PHASE_TO_OUTCOME.get(phase)
 
     @staticmethod
     def _extract_summary(status_dict: dict[str, Any]) -> str:
         """Extract a human-readable summary from analysis results.
 
-        Degraded fallback: only called when ProposalAmender.amend() fails.
-        Reads only condition messages, ignoring proposal_results and child
+        Degraded fallback: only called when AgenticRunAmender.amend() fails.
+        Reads only condition messages, ignoring agentic_run_results and child
         Result CRs (analysis/execution/verification). The full rich summary
-        is built by ProposalAmender.build_summary() in the happy path.
+        is built by AgenticRunAmender.build_summary() in the happy path.
         """
         conditions = status_dict.get("conditions", [])
         messages = [c["message"] for c in conditions if c.get("message")]
         return "; ".join(messages) if messages else "No summary available"
+
+
+# Deprecated alias for backward compatibility
+ProposalDriver = AgenticRunDriver
