--- tests/integration/system-config-agents-proposal.yaml
@@ -1,28 +1,0 @@
-logging:
-  source_level: DEBUG
-
-core:
-  max_threads: 1
-  fail_on_invalid_data: true
-  skip_on_failure: false
-
-agents:
-  enabled: true
-  default:
-    agent:
-      - proposal_agent
-  proposal_agent:
-    type: proposal
-    namespace: lightspeed-evaluation-test
-    auto_approve: true
-    cleanup_proposals: false
-    timeout: 1200
-    poll_interval: 5
-
-storage:
-  - type: file
-    output_dir: ./eval_output
-    enabled_outputs: [csv, json]
-
-environment:
-  LITELLM_LOG: ERROR
